+++ src/main/java/com/example/menu/controller/TableCheckBillController.java
@@ -2,8 +2,10 @@ package com.example.menu.controller;
 
 import com.example.menu.api.APIResponse;
 import com.example.menu.entities.TableCheckBill;
+import com.example.menu.entities.UserManager;
 import com.example.menu.repository.TableCheckBillRepository;
 import com.example.menu.service.TableCheckBillService;
+import com.example.menu.util.ContextUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -20,18 +22,27 @@ public class TableCheckBillController {
     @Autowired
     private TableCheckBillService tableCheckBillService;
 
+    @Autowired
+    private ContextUtil contextUtil;
+
 
     @PostMapping("/delete/{tableCheckBillId}")
     public Object deleteTableCheckBill(@PathVariable int tableCheckBillId) {
         APIResponse response = new APIResponse();
-        Optional<TableCheckBill> checkId = tableCheckBillRepository.findById(tableCheckBillId);
-        if (checkId.isPresent()){
-            tableCheckBillService.deleteTableCheckBill(tableCheckBillId);
-            response.setStatus(1);
-            response.setMessage("delete success");
-        }else{
+        Optional<UserManager> optUserManager = contextUtil.getUserDataFromContext(); /// use token for pull user data.
+        if(optUserManager.isPresent()){
+            Optional<TableCheckBill> checkId = tableCheckBillRepository.findById(tableCheckBillId);
+            if (checkId.isPresent()){
+                tableCheckBillService.deleteTableCheckBill(tableCheckBillId);
+                response.setStatus(1);
+                response.setMessage("delete success");
+            }else{
+                response.setStatus(0);
+                response.setMessage("can't delete");
+            }
+        }else {
             response.setStatus(0);
-            response.setMessage("can't delete");
+            response.setMessage("No UserManager");
         }
         return response;
     }
@@ -39,8 +50,15 @@ public class TableCheckBillController {
     @GetMapping("/list")
     public Object listTableCheckBill() {
         APIResponse response = new APIResponse();
-        List<TableCheckBill> listTableCheckBill = tableCheckBillService.listTableCheckBill();
-        response.setData(listTableCheckBill);
+        Optional<UserManager> optUserManager = contextUtil.getUserDataFromContext(); /// use token for pull user data.
+        if(optUserManager.isPresent()){
+            List<TableCheckBill> listTableCheckBill = tableCheckBillService.listTableCheckBill(optUserManager.get().getManagerId());
+            response.setStatus(1);
+            response.setData(listTableCheckBill);
+        }else {
+            response.setStatus(0);
+            response.setMessage("No UserManager");
+        }
         return response;
     }
